+++ src/main/java/xyz/spoonmap/server/config/jpa/QueryDslConfig.java
@@ -1,18 +1,10 @@
 package xyz.spoonmap.server.config.jpa;
 
-import com.p6spy.engine.common.P6Util;
-import com.p6spy.engine.logging.Category;
-import com.p6spy.engine.spy.P6SpyOptions;
-import com.p6spy.engine.spy.appender.MessageFormattingStrategy;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.util.Locale;
-import javax.annotation.PostConstruct;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import org.hibernate.engine.jdbc.internal.FormatStyle;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
 
 @Configuration
 public class QueryDslConfig {
